infra/vultr/requirements.txt
python-dotenv==1.1.0
boto3==1.37.33
pymongo==4.12.0
mongoengine==0.29.1
